[HADRN-1820] Enable truststore for equinix inventory (#1)
This fixes and issue experianced on experian laptops where their SSL
Proxy certificate wasn't being trusted by Python. Using the truststore
package we are able to use the systems (mac/windows) certificate chain
instead of the one packaged by python, this allows this module to
continue working.

NOTE: This is designed in such a way that it will only enable if
truststore is installed so hopefully it will not create issues on others
systems that do not require this feature.

Co-authored-by: Scott Godbold <scgodbold@gmail.com>